add support for MariaDB

+++ install.sh
@@ -1,4 +1,9 @@
 #!/bin/bash 
+if [ $# = 0 ]; then
+	sql='NO'
+else
+	sql=$1
+fi
 echo								#
 echo " "							#
 echo "Installing the SGP 2D live tracking interface ...." 	#
@@ -32,16 +37,19 @@ echo "=================================================="	#
 echo " "							#
 echo								#
 cd /var/www/html/main						#
-sudo apt-get install -y mariadb-server mariadb-client		#
-sudo apt-get install -y tasksel  				#
-sudo tasksel install -y lamp-server                             #
-#sudo apt policy mysql-server					#
-#sudo apt install mysql-server=5.7.32-1ubuntu18.04		#
-#sudo apt install mysql-client=5.7.32-1ubuntu18.04		#
+sudo apt install -y mariadb-server mariadb-client		#
+sudo apt install -y libmariadb-dev				#
+if [ $sql = 'MySQL' ]			
+then		
+	sudo apt-get install -y tasksel  			#
+	sudo apt policy mysql-server				#
+	sudo apt install mysql-server=5.7.32-1ubuntu18.04	#
+	sudo apt install mysql-client=5.7.32-1ubuntu18.04	#
+fi								#
+sudo tasksel install lamp-server                                #
 sudo apt-get install -y percona-toolkit				#
 sudo apt-get install -y sqlite3					#
 sudo apt-get install -y python3-dev python3-pip 		#
-sudo apt-get install -y python-mysqld  				#
 sudo apt-get install -y figlet inetutils-*			#
 sudo apt-get install -y dos2unix libarchive-dev	 autoconf mc	#
 sudo apt-get install -y pkg-config git	mutt npm nodejs vim	# 
@@ -49,14 +57,12 @@ git config --global user.email "acasadoalonso@gmail.com"        #
 git config --global user.name "Angel Casado"                    #
 sudo apt-get install -y apache2 php 				#
 sudo apt-get install -y php-sqlite3 php-mysql php-cli 		#
-sudo apt-get install -y php-mcrypt 				#
-sudo apt-get install -y php-mbstring php-gettext php-json	#
-sudo apt-get install -y php7.2					#
+sudo apt-get install -y php-mbstring php-json			#
+sudo apt-get install -y php7.4					#
 sudo apt-get install -y ntpdate					#
 sudo apt-get install -y ssmtp					#
-sudo apt-get install -y at sshpass minicomm 			#
+sudo apt-get install -y at sshpass minicom 			#
 sudo a2enmod rewrite						#
-sudo phpenmod mcrypt						#
 sudo phpenmod mbstring						#
 sudo cat /etc/apache2/apache2.conf html.dir 	>>temp.conf	#
 sudo echo "ServerName APRSLOG " >>temp.conf			#
@@ -76,10 +82,15 @@ sudo -H python3 -m pip install tqdm psutil 			#
 sudo -H python3 -m pip install ttn               		#
 sudo -H python3 -m pip install pyserial 			#
 sudo -H python3 -m pip install eciespy pycryptodome rsa         #
+sudo -H python3 -m pip install mariadb               		#
+if [ $sql = 'MySQL' ]					
+then	
+	sudo -H pip3 uninstall mysqlclient			#
+fi
 sudo apt-get install -y libmysqlclient-dev 			#
-sudo -H pip3 uninstall mysqlclient				#
-sudo -H pip3 install --no-binary mysqlclient mysqlclient	#
+sudo -H pip3 install --no-binary mysqlclient mysqlclient 	#
 cd /var/www/html/						#
+sudo npm install -g npm 
 sudo npm install websocket socket.io request parsejson	ini	#
 sudo npm install forever -g 					#
 if [ ! -d /etc/local ]						#
@@ -112,19 +123,28 @@ cd /var/www/html/main						#
 echo "Running msqladmin .... assign root password ... "		#
 sudo mysqladmin -u root password ogn				#
 echo "Create the APRSogn login-path: Type assigned password"	#
-mysql_config_editor set --login-path=APRSogn --user=ogn --password
+if [ $sql = 'MySQL' ]	
+then			
+	mysql_config_editor set --login-path=APRSogn --user=ogn --password
+fi
 echo "Create user ogn ..."					#
 sudo mysql  <doc/adduser.sql					#
 echo "Create database APRSLOG ..."				#
-echo "CREATE DATABASE APRSLOG" | mysql --login-path=APRSogn	#
-mysql --login-path=APRSogn --database APRSLOG < APRSLOG.template.sql #
+if [ $sql = 'MySQL' ]			
+then								#
+	echo "CREATE DATABASE APRSLOG" | mysql --login-path=APRSogn	#
+else
+	echo "CREATE DATABASE APRSLOG" | mysql -u ogn -pogn	
+fi
+if [ $sql = 'MySQL' ]			
+then								#
+    mysql --login-path=APRSogn --database APRSLOG < APRSLOG.template.sql #
+else
+    mysql -u ogn -pogn --database APRSLOG < APRSLOG.template.sql  #
+fi
 echo								#
 echo "Optional steps ... "					#
 echo								#
-#sudo apt-get install -y libsqlite3-dev ruby-dev			#
-#echo "Install MailCatcher."					#
-#sudo gem install mailcatcher					#
-#mailcatcher --http-ip=0.0.0.0					#
 cd sh	 							#
 crontab <crontab.data						#
 crontab -l 							#
@@ -151,6 +171,9 @@ fi								#
 cd								#
 sudo apt-get install percona-toolkit				#
 sudo dpkg-reconfigure tzdata					#
+echo ""								#
+echo "========================================================================================================"	#
+echo ""								#
 sudo apt-get -y dist-upgrade					#
 sudo apt-get -y autoremove					#
 cp /var/www/html/main/doc/aliases .bash_aliases			#
@@ -165,7 +188,6 @@ echo "Check the placement of the RootDocument on APACHE2 ... needs to be /var/ww
 echo "If running in Windows under Virtual Box, run dos2unix on /var/www/html  main  src		"		#
 echo "Install phpmyadmin if needed !!!                                                           "              #
 echo "========================================================================================================"	#
-echo ""								#
 bash
 alias
